FIX: gems path take 3

+++ lib/amazon_affiliate.rb
@@ -172,6 +172,8 @@ module AmazonAffiliate
 
     gems_path = File.expand_path("../gems/#{dev_version}", __dir__)
 
+    return if !File.exists?(gems_path)
+
     FileUtils.mv gems_path, gems_path.gsub(dev_version, RUBY_VERSION)
   end
 end
